some refactoring of tooling code, added RCGossipReceivedEvent
