+++ llarp/CMakeLists.txt
@@ -198,8 +198,10 @@ set(LIB_SRC
   service/session.cpp
   service/tag_lookup_job.cpp
   service/tag.cpp
-  tooling/dht_event.cpp
   tooling/router_event.cpp
+  tooling/dht_event.cpp
+  tooling/path_event.cpp
+  tooling/rc_event.cpp
 )
   
 if(TRACY_ROOT)
