+++ llarp/messages/relay_status.cpp
@@ -10,7 +10,7 @@
 #include <util/logging/logger.hpp>
 #include <util/meta/memfn.hpp>
 #include <util/thread/logic.hpp>
-#include <tooling/router_event.hpp>
+#include <tooling/path_event.hpp>
 
 #include <functional>
 #include <utility>
