+++ llarp/router/router.cpp
@@ -58,7 +58,11 @@ namespace llarp
       , _dht(llarp_dht_context_new(this))
       , inbound_link_msg_parser(this)
       , _hiddenServiceContext(this)
+#ifdef LOKINET_HIVE
+      , _randomStartDelay(std::chrono::milliseconds((llarp::randint() % 1250) + 1000))
+#else
       , _randomStartDelay(std::chrono::seconds((llarp::randint() % 30) + 10))
+#endif
   {
     m_keyManager = std::make_shared< KeyManager >();
