+++ test/CMakeLists.txt
@@ -1,10 +1,11 @@
 if (WITH_HIVE)
+  add_custom_target(hive_build DEPENDS ${STATIC_LIB} pyllarp)
   add_custom_target(hive ${CMAKE_COMMAND} -E
       env PYTHONPATH="$ENV{PYTHONPATH}:${CMAKE_BINARY_DIR}/pybind"
       ${PYTHON_EXECUTABLE} -m pytest
       ${CMAKE_CURRENT_SOURCE_DIR}/hive
       DEPENDS
-      ${STATIC_LIB} pyllarp)
+      hive_build)
 endif()
 
 set(GTEST_EXE testAll)
